Move long script variables to actual script files
