+++ gt/tools/batch_processor/__init__.py
@@ -10,7 +10,7 @@ logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
 
 # Tool Version
-__version_tuple__ = (0, 1, 0)
+__version_tuple__ = (1, 0, 0)
 __version_suffix__ = ""
 __version__ = ".".join(str(number) for number in __version_tuple__) + __version_suffix__
