+++ gt/tools/batch_processor/tasks/task_external_blender.py
@@ -2,130 +2,28 @@
 Batch Processor Blender Script Task
 """
 
+
 from gt.tools.batch_processor import batch_processor_constants as constants
+from gt.tools.batch_processor.tasks import task_utils
 from gt.tools.batch_processor import batch_processor_task_base as task_base
-import gt.ui.resource_library as ui_res_lib
 from gt.tools.batch_processor.tasks.task_external_mobu import TaskMotionBuilderScript
-from gt.tools.batch_processor.tasks.task_external_mobu import parse_command_arguments
+from gt.tools.batch_processor.tasks.task_external_script import TaskExternalScript
+from gt.tools.batch_processor.tasks.task_external_script import parse_command_arguments
 from gt.tools.batch_processor.tasks.task_python_script import SCRIPT_MODE_INLINE
 from gt.tools.batch_processor.tasks.task_python_script import TaskPythonScript
 import glob
 import os
 import sys
+import gt.ui.resource_library as ui_res_lib
+
 
 
 DEFAULT_BLENDER_ARGUMENTS = "--background"
 DEFAULT_BLENDER_SCRIPT_FLAG = "--python"
-DEFAULT_BLENDER_INLINE_SCRIPT = r'''"""
-Blender batch example.
-
-Arguments passed by the batch processor when Pass Task Args is enabled:
-    --input: Source file.
-        Example: C:/project/01_input/character.fbx
-    --output: Expected output file.
-        Example: C:/project/02_tasks/01_blender/character.fbx
-    --project: Batch project file.
-        Example: C:/project/Batch_process_project.batch
-    --project-dir: Folder containing the batch project.
-        Example: C:/project
-    --task: Task display name.
-        Example: Blender
-    --task-id: Stable task id from the .batch file.
-        Example: 2f35ec56-8e64-4df7-9a30-612af8d4f2db
-    --context: JSON file with the same batch context data.
-        Example: C:/Users/name/AppData/Local/Temp/blender_context_ab12.json
-    --env-var: Repeated project environment variable pair.
-        Example: --env-var output-dir=03_output
-
-The task also sets BATCH_INPUT, BATCH_OUTPUT, BATCH_PROJECT,
-BATCH_PROJECT_DIR, BATCH_TASK, BATCH_TASK_ID, and BATCH_CONTEXT as process
-environment variables.
-"""
-
-import argparse
-import json
-import os
-import sys
-
-import bpy
+DEFAULT_BLENDER_INLINE_SCRIPT = task_utils.load_script("script_inline_external_blender.py")
 
 
-def parse_args():
-    """Parses arguments after Blender's -- separator."""
-    parser = argparse.ArgumentParser(add_help=False)
-    parser.add_argument("--input", default="")
-    parser.add_argument("--output", default="")
-    parser.add_argument("--project", default="")
-    parser.add_argument("--project-dir", default="")
-    parser.add_argument("--task", default="")
-    parser.add_argument("--task-id", default="")
-    parser.add_argument("--context", default="")
-    parser.add_argument("--env-var", action="append", default=[])
-    argv = sys.argv
-    if "--" in argv:
-        argv = argv[argv.index("--") + 1:]
-    else:
-        argv = []
-    args, _ = parser.parse_known_args(argv)
-    return args
-
-
-def load_context(context_path):
-    """Loads the optional batch context JSON file."""
-    if context_path and os.path.isfile(context_path):
-        with open(context_path, "r", encoding="utf-8") as context_file:
-            return json.load(context_file)
-    return {}
-
-
-def open_input_file(file_path):
-    """Opens a Blend file or imports a supported interchange file."""
-    extension = os.path.splitext(file_path)[1].lower()
-    if extension == ".blend":
-        bpy.ops.wm.open_mainfile(filepath=file_path)
-        return
-    bpy.ops.wm.read_factory_settings(use_empty=True)
-    if extension == ".fbx":
-        bpy.ops.import_scene.fbx(filepath=file_path)
-    elif extension == ".obj":
-        bpy.ops.wm.obj_import(filepath=file_path)
-    elif extension in [".gltf", ".glb"]:
-        bpy.ops.import_scene.gltf(filepath=file_path)
-    else:
-        raise RuntimeError("Unsupported Blender input extension: {0}".format(extension))
-
-
-def export_fbx(file_path):
-    """Exports the current Blender scene as FBX."""
-    output_path = os.path.splitext(file_path)[0] + ".fbx"
-    output_dir = os.path.dirname(output_path)
-    if output_dir and not os.path.isdir(output_dir):
-        os.makedirs(output_dir)
-    bpy.ops.export_scene.fbx(filepath=output_path)
-    return output_path
-
-
-def main():
-    """Runs the Blender batch example."""
-    args = parse_args()
-    context_path = args.context or os.environ.get("BATCH_CONTEXT", "")
-    context = load_context(context_path)
-    input_path = args.input or os.environ.get("BATCH_INPUT") or context.get("current_path") or context.get("source_path")
-    output_path = args.output or os.environ.get("BATCH_OUTPUT") or context.get("output_path")
-    if not input_path or not output_path:
-        raise RuntimeError("Input and output paths are required.")
-    print("Input Path: {0}".format(input_path))
-    print("Output Path: {0}".format(output_path))
-    open_input_file(input_path)
-    exported_path = export_fbx(output_path)
-    print("Exported FBX: {0}".format(exported_path))
-
-
-main()
-'''
-
-
-class TaskBlenderScript(TaskMotionBuilderScript):
+class TaskBlenderScript(TaskExternalScript):
     """Task that runs Blender Python scripts as an external process."""
 
     task_type = constants.TaskType.BLENDER_SCRIPT
@@ -144,7 +42,7 @@ class TaskBlenderScript(TaskMotionBuilderScript):
         Returns:
             dict: Default settings.
         """
-        settings = TaskPythonScript.get_default_settings(self)
+        settings = super().get_default_settings()
         settings.update(
             {
                 "source_path": "{previous-task-path}",
@@ -170,32 +68,7 @@ class TaskBlenderScript(TaskMotionBuilderScript):
         )
         return settings
 
-    def validate(self, project):
-        """Validates Blender script settings.
-
-        Args:
-            project (BatchProcessorModel): Project containing this task.
-
-        Returns:
-            ValidationResult: Collected validation result.
-        """
-        result = task_base.ValidationResult()
-        executable_path = self.resolve_motionbuilder_executable(project)
-        if not executable_path:
-            result.add_error("Blender executable path cannot be empty.")
-        elif not os.path.isfile(executable_path):
-            result.add_error("Blender executable does not exist: {0}".format(executable_path))
-        if self.is_batch_mode():
-            result.extend(self.validate_batch_settings(project))
-        elif self.is_external_file_mode():
-            result.extend(self.validate_external_file_settings(project))
-        else:
-            result.extend(self.validate_single_script_settings(project))
-        if not self.settings.get("wait_for_completion") and self.settings.get("require_output_file"):
-            result.add_warning("Require Output is ignored when Wait For Completion is disabled.")
-        return result
-
-    def build_motionbuilder_command(self, executable_path, script_path, work_item, output_path, project, context_path):
+    def build_external_command(self, executable_path, script_path, work_item, output_path, project, context_path):
         """Builds the Blender process command.
 
         Args:
@@ -241,6 +114,7 @@ class TaskBlenderScript(TaskMotionBuilderScript):
                 command.extend(["--env-var", "{0}={1}".format(key, value)])
         return command
 
+
     def build_process_log_path(self, output_path, step_output_dir, script_path, project=None):
         """Builds a process log path for Blender output.
 
@@ -265,6 +139,7 @@ class TaskBlenderScript(TaskMotionBuilderScript):
         )
         return task_base.normalize_path(os.path.join(target_dir, file_name))
 
+
     def finalize_output(self, work_item, output_path):
         """Resolves Blender output without requiring a specific extension.
 
@@ -286,7 +161,8 @@ class TaskBlenderScript(TaskMotionBuilderScript):
             raise RuntimeError(f"Blender script did not create an output named '{output_name}' in: {output_dir}")
         return output_path
 
-    def resolve_motionbuilder_executable(self, project=None):
+
+    def resolve_executable(self, project=None):
         """Resolves the configured Blender executable path.
 
         Args:
